Refactoring TransferPanel.
